api/requirements.txt
python-dotenv
requests
gunicorn
fastapi
pyzbar
uvicorn
python-multipart
pillow
memory_profiler
pytest
pytest_cov
flake8
fastapi-healthcheck-flagman
python-jose
passlib
sqlalchemy
pydantic
pydantic[email]
alembic
psycopg2-binary